Remove "chai" / Fix Lint / Improve PrettierIgnore / Rename folder "tests/function" to "tests/e2e" (see tutorial, examples)
