+++ tutorial/lesson-4/package.json
@@ -45,7 +45,6 @@
     "aws-sdk": "2.860.0",
     "aws-xray-sdk-core": "3.2.0",
     "babel-jest": "27.0.2",
-    "chai": "4.3.3",
     "isomorphic-fetch": "3.0.0",
     "jest": "27.0.3",
     "react-test-renderer": "17.0.2",
